+++ backend/src/main/java/com/agrupae/application/exception/course/TargetUserNotEnrolled.java
@@ -1,0 +1,9 @@
+package com.agrupae.application.exception.course;
+
+import com.agrupae.application.exception.ApplicationException;
+
+public class TargetUserNotEnrolled extends ApplicationException{
+    public TargetUserNotEnrolled(){
+        super("It's only possible to transfer leadership of a course to a member.");
+    }
+}